Added sub directories to unit tests. Updated namespace for unit tests moved. All tests passing
