+++ ConwaysGameOfLife/Cell/Orientations/Orientation.cs
