+++ ConwaysGameOfLife/Cell/Strategies/BottomLeftCornerStrategy.cs
@@ -1,14 +1,14 @@
 using System.Collections.Generic;
-using ConwaysGameOfLife.Model;
+using ConwaysGameOfLife.States;
 
-namespace ConwaysGameOfLife.Orientations.Strategies
+namespace ConwaysGameOfLife.Strategies
 {
-    public static class LeftSideStrategy
+    public static class BottomLeftCornerStrategy
     {
         public static List<State> GetNeighbourCellsState(Cell cell, Cell[,] universeGrid, int universeDimensions)
         {
             var neighbour = new Neighbour(cell, universeDimensions);
-            var rows = new List<int> {neighbour.PreviousRow, neighbour.NextRow};
+            var rows = new List<int> {neighbour.PreviousRow, Constants.FirstRowOrColumn};
             var columns = new List<int> {neighbour.LastRowOrColumn, neighbour.NextColumn};
             var neighbourCellsState = neighbour.GetNeighbourCellsState(universeGrid, rows, columns);
             return neighbourCellsState;
