+++ ConwaysGameOfLife/Cell/Strategies/LeftSideStrategy.cs
@@ -1,15 +1,15 @@
 using System.Collections.Generic;
-using ConwaysGameOfLife.Model;
+using ConwaysGameOfLife.States;
 
-namespace ConwaysGameOfLife.Orientations.Strategies
+namespace ConwaysGameOfLife.Strategies
 {
-    public static class TopRightCornerStrategy
+    public static class LeftSideStrategy
     {
         public static List<State> GetNeighbourCellsState(Cell cell, Cell[,] universeGrid, int universeDimensions)
         {
             var neighbour = new Neighbour(cell, universeDimensions);
-            var rows = new List<int> {neighbour.LastRowOrColumn, neighbour.NextRow};
-            var columns = new List<int> {Constants.FirstRowOrColumn, neighbour.PreviousColumn};
+            var rows = new List<int> {neighbour.PreviousRow, neighbour.NextRow};
+            var columns = new List<int> {neighbour.LastRowOrColumn, neighbour.NextColumn};
             var neighbourCellsState = neighbour.GetNeighbourCellsState(universeGrid, rows, columns);
             return neighbourCellsState;
         }
