+++ ConwaysGameOfLife/Cell/Strategies/TopRightCornerStrategy.cs
@@ -1,15 +1,15 @@
 using System.Collections.Generic;
-using ConwaysGameOfLife.Model;
+using ConwaysGameOfLife.States;
 
-namespace ConwaysGameOfLife.Orientations.Strategies
+namespace ConwaysGameOfLife.Strategies
 {
-    public static class BottomLeftCornerStrategy
+    public static class TopRightCornerStrategy
     {
         public static List<State> GetNeighbourCellsState(Cell cell, Cell[,] universeGrid, int universeDimensions)
         {
             var neighbour = new Neighbour(cell, universeDimensions);
-            var rows = new List<int> {neighbour.PreviousRow, Constants.FirstRowOrColumn};
-            var columns = new List<int> {neighbour.LastRowOrColumn, neighbour.NextColumn};
+            var rows = new List<int> {neighbour.LastRowOrColumn, neighbour.NextRow};
+            var columns = new List<int> {Constants.FirstRowOrColumn, neighbour.PreviousColumn};
             var neighbourCellsState = neighbour.GetNeighbourCellsState(universeGrid, rows, columns);
             return neighbourCellsState;
         }
