+++ ConwaysGameOfLifeTest/CellTests/CoordinateTest.cs
@@ -1,7 +1,7 @@
-using ConwaysGameOfLife.Model;
+using ConwaysGameOfLife;
 using Xunit;
 
-namespace ConwaysGameOfLifeTest
+namespace ConwaysGameOfLifeTest.CellTests
 {
     public class CoordinateTest
     {
@@ -12,11 +12,10 @@ namespace ConwaysGameOfLifeTest
         public void Try_Parse_Should_Return_True_If_Input_For_Coordinate_Is_Valid(string input, int xPosition, int yPosition)
         {
             //arrange
-            Coordinate coordinate = new Coordinate(xPosition, yPosition);
             var expectedCoordinate = new Coordinate(xPosition, yPosition);
 
             //act
-            var actualCoordinate = Coordinate.TryParse(input, out coordinate);
+            var actualCoordinate = Coordinate.TryParse(input, out var coordinate);
 
             //assert
             Assert.True(actualCoordinate);
@@ -33,10 +32,9 @@ namespace ConwaysGameOfLifeTest
         public void Try_Parse_Should_Return_False_If_Input_For_Coordinate_Is_Invalid(string input)
         {
             //arrange
-            Coordinate coordinate = null;
 
             //act
-            var actualCoordinate = Coordinate.TryParse(input, out coordinate);
+            var actualCoordinate = Coordinate.TryParse(input, out _);
 
             //assert
             Assert.False(actualCoordinate);
