+++ ConwaysGameOfLifeTest/SeedTests/SeedCreatorTest.cs
@@ -2,11 +2,12 @@ using System;
 using System.Collections.Generic;
 using ConwaysGameOfLife;
 using ConwaysGameOfLife.Console;
-using ConwaysGameOfLife.Model;
+using ConwaysGameOfLife.Seed;
+using ConwaysGameOfLife.States;
 using Moq;
 using Xunit;
 
-namespace ConwaysGameOfLifeTest
+namespace ConwaysGameOfLifeTest.SeedTests
 {
     public class SeedCreatorTest
     {
